NSAC: add lexical unit type CALC.
NSAC 2.0 was somewhat different to 1.0, but to ease the transition it tried to keep the same lexical types (with the exception of var() which is a different beast). For NSAC 2.1, now is time to handle calc() as its own lexical type.

This commit applies changes to the NSAC implementation and the CSSOM as well.