NSAC impl.: change how escaped backslashes are handled.
Escaped backslahes were being processed like normal characters, and then
ParseHelper.escapeBackslash() had to make certain assumptions. That was
error-prone (especially for users of escapeBackslash()); now it is handled
with a bit more overhead, but the logic is easier to follow (and results in
a few buggy edge cases being handled correctly).