NSAC: ERRATUM: document the actual getStringValue()/getParameters() behaviour of 'var()' values in LexicalUnit.
The LexicalUnit javadocs for VAR said that getStringValue() returned the custom property name and getParameters() the fallback.

This is not what the actual behaviour is: getParameters() gives the property name and, eventually, the comma and the fallback.
Modify the javadoc instead of changing the behaviour.

(cherry picked from commit acb237b891bae9445747b63367c9d7c1f975d838)