NSAC/CSSOM: use 'matches(CSSValueSyntax)' to implement registered custom properties.
This is a follow-up to commit [b67e1aad]. Syntax verification of initial values
is done through 'matches', and this method accepts values with non-absolute units.
This could create some problems, as initial values are intended to be absolute,
so if there is user interest I could implement a 'isValidInitial()' or anything
like that, and use it to check initial values.

NSAC API CHANGE: new lexicalProperty() method was added to CSSHandler.