DOM: add `rebuildCascade()` to `CSSDocument`.
In some cases, the library does not detect that a property value has changed,
and keeps using an outdated cascade. With the new `rebuildCascade()` method,
developers can now make sure that the cascade is rebuilt once they are done
modifying the properties.

The interface provides a default implementation so it can be used with older
versions of css4j-dom4j.