NSAC: require `Flag.IEVALUES` to parse IE expressions
This implies a stricter parsing of `calc()` expressions involving a plus sign.

Previously, the library would accept values like `calc(2+2)` while correctly
serializing them as `calc(2 + 2)`, because they are unambiguous and some legacy
IE expressions allow that syntax, but that gets in the way of the people that
use the library for error detection.