The advanced `attr()` support now follows the current Values Level 5 specification
Should be compatible with the `attr()` which is shipped with Google Chrome 133.
The old `CSSAttrValue` and `AttrValue` classes are removed, as the API is
incompatible with the new specification that is being implemented by browsers.

Also, the handling of registered custom properties is now closer to the Google
Chrome behaviour: registered initial values take precedence over the fallbacks.
On `var()` circularities (and other DoS attacks), it is no longer attempted to
use the supplied property fallbacks.