CSS Nesting (II): object model
The code is intended to be compatible with the latest specification, which was
implemented by the major web browsers in late 2024 (Chrome 130, Firefox 132 and
Safari 18.2).

The style computation does not attempt to handle the case where the relative
selector is not a compound one and the scope selector is not used explicitly, as
the specification does not clearly tell what to do. Future follow-up commits
should address that.